refactor: component pages로 이동
